+++ CHANGELOG.md
@@ -5,7 +5,12 @@ All notable changes to the **daily-python** SDK will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.12.1] - 2024-10-29
+## [Unreleased]
+
+### Added
+
+- Added support for `dialin-connected`, `dialin-stopped`, `dialin-warning`, and
+  `dialin-error` events.
 
 ### Fixed